+++ app/overrides/add_welcome_bottom_content.rb
@@ -1,5 +1,0 @@
-Deface::Override.new virtual_path: 'welcome/index',
-                     name:         'add-welcome-bottom-content',
-                     insert_after: 'div.splitcontentright',
-                     original:     'dd470844bcaa4d7c9dc66e70e6c0c843d42969bf',
-                     partial:      'welcome/overview_bottom'