+++ app/overrides/add_welcome_top_content.rb
@@ -1,5 +1,0 @@
-Deface::Override.new virtual_path:  'welcome/index',
-                     name:          'add-welcome-top-content',
-                     insert_before: 'div.splitcontentleft',
-                     original:      'e7de0a2e88c5ccb4d1feb7abac239e4b669babed',
-                     partial:       'welcome/overview_top'