+++ app/overrides/remove_welcome_news.rb
@@ -1,5 +1,0 @@
-Deface::Override.new virtual_path: 'welcome/index',
-                     name:         'remove-welcome-news',
-                     replace:      'div.news',
-                     original:     '163f5df8f0cb2d5009d7f57ad38174ed29201a1a',
-                     partial:      'welcome/remove_news'