+++ app/overrides/users/show.rb
@@ -1,0 +1,5 @@
+Deface::Override.new virtual_path: 'users/show',
+                     name: 'user-show-info-hook',
+                     insert_top: 'div.splitcontentleft ul:first-child',
+                     original: 'aff8d775275e3f33cc45d72b8e2896144be4beff',
+                     partial: 'hooks/view_users_show'