+++ test/functional/wiki_controller_test.rb
@@ -8,6 +8,7 @@ class WikiControllerTest < ActionController::TestCase
            :members,
            :member_roles,
            :trackers,
+           :groups_users,
            :projects_trackers,
            :enabled_modules,
            :issue_statuses,
@@ -259,7 +260,31 @@ class WikiControllerTest < ActionController::TestCase
     assert_response :success
     assert_template 'show'
     assert_select 'a[href=?]', '/issues/2',
-                  text: 'Add ingredients categories #2'
+                  text: '#2 Add ingredients categories'
+  end
+
+  def test_show_issue_with_url
+    @request.session[:user_id] = WIKI_MACRO_USER_ID
+    @page.content.text = '{{issue(http://test.host/issues/2)}}'
+    @page.content.save!
+    get :show, project_id: 1, id: @page_name
+    assert_response :success
+    assert_template 'show'
+    assert_select 'a[href=?]', '/issues/2',
+                  text: '#2 Add ingredients categories'
+    assert_select 'div.issue-macro-comment', 0
+  end
+
+  def test_show_issue_and_comment_with_url
+    @request.session[:user_id] = WIKI_MACRO_USER_ID
+    @page.content.text = '{{issue(http://test.host/issues/2#note-1)}}'
+    @page.content.save!
+    get :show, project_id: 1, id: @page_name
+    assert_response :success
+    assert_template 'show'
+    assert_select 'a[href=?]', '/issues/2',
+                  text: '#2 Add ingredients categories'
+    assert_select 'div.issue-macro-comment'
   end
 
   def test_show_issue_with_id_default
@@ -270,7 +295,7 @@ class WikiControllerTest < ActionController::TestCase
     assert_response :success
     assert_template 'show'
     assert_select 'a[href=?]', '/issues/2',
-                  text: 'Add ingredients categories #2'
+                  text: '#2 Add ingredients categories'
   end
 
   def test_show_user_with_id