+++ test/unit/helpers/global_helper_test.rb
@@ -51,4 +51,32 @@ class GlobalHelperTest < ActionView::TestCase
     html = font_awesome_icon('fa-car', post_text: 'Testing')
     assert_include '</span> Testing', html
   end
+
+  def test_parse_issue_url
+    stubs(:request).returns(stub('original_url' => 'http://redmine.local/issues/1#note-2'))
+
+    assert_equal({ issue_id: nil, comment_id: nil },
+                 parse_issue_url(0, nil))
+    assert_equal({ issue_id: nil, comment_id: nil },
+                 parse_issue_url('', nil))
+    assert_equal({ issue_id: nil, comment_id: nil },
+                 parse_issue_url('http://localhost/issues/23', nil))
+    assert_equal({ issue_id: '23', comment_id: nil },
+                 parse_issue_url('http://redmine.local/issues/23', nil))
+    assert_equal({ issue_id: '23', comment_id: 2 },
+                 parse_issue_url('http://redmine.local/issues/23#note-2', nil))
+    assert_equal({ issue_id: '23', comment_id: 2 },
+                 parse_issue_url('http://redmine.local/issues/issues/23/edit#note-2', nil))
+  end
+
+  def test_render_issue_macro_link
+    stubs(:request).returns(stub('original_url' => 'http://redmine.local/issues/1#note-2'))
+
+    issue = Issue.find(1)
+
+    assert_match %r{/issues/1}, render_issue_macro_link(issue, 'Sample subject')
+    assert_no_match(/Journal notes/, render_issue_macro_link(issue, 'Sample subject'))
+    assert_match(/Journal notes/, render_issue_macro_link(issue, 'Sample subject', 1))
+    assert_match %r{N/A}, render_issue_macro_link(issue, 'Sample subject', 100)
+  end
 end