+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Redmine 3.4 bug fixes
 - Helper function fa_icon renamed to font_awesome_icon because of conflicts with redmine_bootstrap_kit
+- Cleanups deface overwrites
+- add hook for user show
 
 2.0.4
 +++++